Move SapSystemDetails to pages
