+++ assets/js/pages/InstanceOverview/InstanceOverview.jsx
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 
 import { DATABASE_TYPE } from '@lib/model/sapSystems';
 import HealthIcon from '@pages/HealthIcon';
-import { Features } from '@components/SapSystemDetails';
+import { Features } from '@pages/SapSystemDetails';
 import HostLink from '@components/HostLink';
 import ClusterLink from '@pages/ClusterDetails/ClusterLink';
 import Pill from '@common/Pill';
