+++ assets/js/pages/SapSystemDetails/GenericSystemDetails.jsx
