+++ assets/js/pages/SapSystemDetails/GenericSystemDetails.test.jsx
