+++ assets/js/pages/SapSystemDetails/SapSystemDetails.stories.jsx
