+++ assets/js/pages/SapSystemDetails/tableConfigs.jsx
