+++ assets/js/trento.jsx
@@ -32,7 +32,7 @@ import { networkClient } from '@lib/network';
 import Guard from '@components/Guard';
 import { TARGET_CLUSTER, TARGET_HOST } from '@lib/model';
 import DatabaseDetails from './pages/DatabaseDetails';
-import SapSystemDetails from './components/SapSystemDetails/SapSystemDetails';
+import SapSystemDetails from './pages/SapSystemDetails/SapSystemDetails';
 
 import { store } from './state';
